# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para reservas, servicios, combos y administracion de Atrevida Fit.",
        "title": "Atrevida Fit - Agenda API",
        "contact": {},
        "version": "v0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/bd/combos/servicios": {
            "post": {
                "description": "Crea un item dentro de combo_servicios. Body: combo_id ID del combo padre (requerido), servicio_id (opcional si se envia servicio_texto), servicio_texto (opcional si se envia servicio_id), tiempo HH:MM (opcional), costo (opcional), sesiones entero positivo default 1 (opcional), orden posicion (opcional). Response: id (int ID del item creado).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Combo Servicios BD"
                ],
                "summary": "Crear servicio de combo",
                "parameters": [
                    {
                        "description": "Datos del servicio del combo",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.crearComboServicioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.idResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: combo_id requerido, sesiones debe ser positivo, debe enviar servicio_id o servicio_texto",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Combo no encontrado o inactivo",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/bd/combos/servicios/{id}": {
            "get": {
                "description": "Devuelve un item de combo_servicios por su ID. Param: id (requerido, path). Response: servicio (ComboServicioDetallePG con: id, combo_id, combo_nombre, servicio_id, servicio_texto, servicio_nombre, tiempo HH:MM, costo, sesiones, orden).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Combo Servicios BD"
                ],
                "summary": "Obtener servicio de combo por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 15,
                        "description": "ID del item combo_servicios",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.comboServicioItemResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Item no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un item de combo_servicios por su ID. Param: id (requerido, path). Response: mensaje string.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Combo Servicios BD"
                ],
                "summary": "Eliminar servicio de combo",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 15,
                        "description": "ID del item combo_servicios",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.messageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Item no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualiza campos de un item de combo_servicios. No permite cambiar el combo padre. Param: id (requerido, path). Body: servicio_id (opcional), servicio_texto (opcional), tiempo HH:MM (opcional), costo (opcional), sesiones (opcional), orden (opcional). Response: mensaje string.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Combo Servicios BD"
                ],
                "summary": "Actualizar servicio de combo",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 15,
                        "description": "ID del item combo_servicios",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar (todos opcionales, al menos uno requerido)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.actualizarComboServicioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.messageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido, sin campos a modificar, sesiones debe ser positivo",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Item no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/bd/combos/{combo_id}/servicios": {
            "get": {
                "description": "Devuelve los items de combo_servicios asociados a un combo activo. Param: combo_id (requerido, path). Response: total (int), combo_id (int), servicios ([]ComboServicioDetallePG con: id, combo_id, combo_nombre, servicio_id, servicio_texto, servicio_nombre, tiempo HH:MM, costo, sesiones, orden).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Combo Servicios BD"
                ],
                "summary": "Listar servicios de un combo",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 12,
                        "description": "ID del combo",
                        "name": "combo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.comboServicioListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: combo_id invalido",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Combo no encontrado o inactivo",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.actualizarComboServicioRequest": {
            "type": "object",
            "properties": {
                "costo": {
                    "type": "number",
                    "example": 250
                },
                "orden": {
                    "type": "integer",
                    "example": 1
                },
                "servicio_id": {
                    "type": "integer",
                    "example": 8
                },
                "servicio_texto": {
                    "type": "string",
                    "example": "Masaje relajante personalizado"
                },
                "sesiones": {
                    "type": "integer",
                    "example": 2
                },
                "tiempo": {
                    "type": "string",
                    "example": "01:00"
                }
            }
        },
...
        "handlers.comboServicioItemResponse": {
            "type": "object",
            "properties": {
                "servicio": {
                    "$ref": "#/definitions/models.ComboServicioDetallePG"
                }
            }
        },
        "handlers.comboServicioListResponse": {
            "type": "object",
            "properties": {
                "combo_id": {
                    "type": "integer",
                    "example": 12
                },
                "servicios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ComboServicioDetallePG"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "handlers.crearComboServicioRequest": {
            "type": "object",
            "required": [
                "combo_id"
            ],
            "properties": {
                "combo_id": {
                    "type": "integer",
                    "example": 12
                },
                "costo": {
                    "type": "number",
                    "example": 250
                },
                "orden": {
                    "type": "integer",
                    "example": 1
                },
                "servicio_id": {
                    "type": "integer",
                    "example": 8
                },
                "servicio_texto": {
                    "type": "string",
                    "example": "Masaje relajante personalizado"
                },
                "sesiones": {
                    "type": "integer",
                    "example": 2
                },
                "tiempo": {
                    "type": "string",
                    "example": "01:00"
                }
            }
        },
...
        "handlers.idResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "handlers.messageResponse": {
            "type": "object",
            "properties": {
                "mensaje": {
                    "type": "string",
                    "example": "operacion realizada correctamente"
                }
            }
        },
...
        "models.ComboServicioDetallePG": {
            "type": "object",
            "properties": {
                "combo_id": {
                    "type": "integer",
                    "example": 12
                },
                "combo_nombre": {
                    "type": "string",
                    "example": "Combo Relax"
                },
                "costo": {
                    "type": "number",
                    "example": 250
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "orden": {
                    "type": "integer",
                    "example": 1
                },
                "servicio_id": {
                    "type": "integer",
                    "example": 8
                },
                "servicio_nombre": {
                    "type": "string",
                    "example": "Masaje relajante personalizado"
                },
                "servicio_texto": {
                    "type": "string",
                    "example": "Masaje relajante personalizado"
                },
                "sesiones": {
                    "type": "integer",
                    "example": 2
                },
                "tiempo": {
                    "type": "string",
                    "example": "01:00"
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "error": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "operacion realizada correctamente"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        }
    }
}